backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
python-multipart
